Crate hopli_lib

Modules

constants

environment_config

This module contains definiation of arguments that specify the environment and networks that a HOPR node runs in.

exports

faucet

This module contains arguments and functions to fund some Ethereum wallets with native tokens and HOPR tokens

identity

This module contains subcommands for hopli identity, This command contains subcommands to read, create or update identity files, providing correct crate::key_pair::PasswordArgs

key_pair

This module contains struct definition, utility functions around private keys, password, and keystores.

methods

This module contains all the methods used for onchain interaction, especially with Safe instance, Mutlicall, and Multisend contracts.

payloads

safe_module

This module contains arguments and functions to manage safe and module. SafeModuleSubcommands defines three subcommands: create, move, and migrate.

utils

This module contains errors produced in this crate

win_prob

This module contains arguments and functions to interact with the Winning Probability contract for a privileged account. It can set the global minimum winning probability and read the current global minimum winning probability. Some sample commands:

Structs

Address

An Ethereum address, 20 bytes in length.