hopli_lib

Module payloads

Structs

UserData

Functions

edge_node_deploy_safe_module_and_maybe_include_node

Deploy a safe and a module, while sending tokens to the safe for single edge node. It’s possible to only deploy a safe and a module without onboarding the node Alternatively, the node will be included in the module after deployment All the “admins” will be the nodes added to the module Returns safe proxy address and module proxy address

edge_node_deploy_safe_module_with_targets_and_nodes_payload

Payload for deploying safe module with target addresses and node addresses This payload assumes the threshold is one. The prediceted safe address can be obtained by calling edge_node_predict_safe_address The predicted module address can be obtained by calling predictModuleAddress_1 on hopr_node_stake_factory providing MULTICALL3_ADDRESS, nonce.into(), safe_address, default_target

edge_node_predict_module_address

edge_node_predict_safe_address

Predict safe address deployed by the edge node

set_winning_probability

Set the global minimum winning probability in the HoprWinningProbabilityOracle contract

transfer_hopr_token_payload

transfer_native_token_payload